Feature/persist thumbnail (#11)
* Check is category file in joiner

* Add callback with multiple ACK and recv batch message

* Call callback with multiple ACK

* PIka log in ERROR

* Add persist method with pass

* Persist data function

* Identation

* Persist country count and data in two files

* Save data in disk and load with checking when started

* Format Dockerfiles

* Fix some prefix

* Mini fix check results

* Add stopping to manual chaos

* Add ranmdon chaos script. Selecting a random killeable and print it

* Print random container kill

* Separate random select function

+++ tp1/common/manual_chaos.py
@@ -4,11 +4,14 @@ import signal
 
 class ManualChaos:
     def __init__(self) -> None:
-        pass
+        self.stop_chaos = False
 
     def chaos(self):
-        print("Break now? Y/n\n")
-        chaos = input()
-        if chaos == "Y":
-            process = current_process()
-            os.kill(process.pid, signal.SIGTERM)
+        if not self.stop_chaos:
+            print("Break up now? Y/n (Use S to stop chaos)\n")
+            chaos = input()
+            if chaos == "Y":
+                process = current_process()
+                os.kill(process.pid, signal.SIGTERM)
+            if chaos == "S":
+                self.stop_chaos = True
